+++ src/main/kotlin/io/github/nchaugen/tabletest/language/TableTestFile.kt
@@ -1,0 +1,10 @@
+package io.github.nchaugen.tabletest.language
+
+import com.intellij.extapi.psi.PsiFileBase
+import com.intellij.psi.FileViewProvider
+
+class TableTestFile(viewProvider: FileViewProvider): PsiFileBase(viewProvider, TableTestLanguage.INSTANCE) {
+    override fun getFileType() = TableTestFileType.INSTANCE
+
+    override fun toString() = "TableTest File"
+}