+++ src/main/kotlin/io/github/nchaugen/tabletest/language/TableTestIcons.kt
@@ -1,0 +1,9 @@
+package io.github.nchaugen.tabletest.language
+
+import com.intellij.openapi.util.IconLoader
+
+class TableTestIcons {
+    companion object {
+        val FILE = IconLoader.getIcon("/icons/table/pagination.svg", TableTestIcons::class.java)
+    }
+}