+++ src/main/kotlin/io/github/nchaugen/tabletest/language/TableTestInjectedFormattingOptionsProvider.kt
@@ -1,0 +1,8 @@
+package io.github.nchaugen.tabletest.language
+
+import com.intellij.formatting.InjectedFormattingOptionsProvider
+import com.intellij.psi.PsiFile
+
+class TableTestInjectedFormattingOptionsProvider: InjectedFormattingOptionsProvider {
+    override fun shouldDelegateToTopLevel(psiFile: PsiFile): Boolean? = psiFile !is TableTestFile
+}