+++ src/main/kotlin/io/github/nchaugen/tabletest/language/TableTestLanguage.kt
@@ -1,0 +1,10 @@
+package io.github.nchaugen.tabletest.language
+
+import com.intellij.lang.Language
+
+class TableTestLanguage: Language("TableTest") {
+    companion object {
+        val INSTANCE = TableTestLanguage()
+    }
+
+}