+++ src/main/kotlin/io/github/nchaugen/tabletest/language/TableTestLexerAdapter.kt
@@ -1,0 +1,5 @@
+package io.github.nchaugen.tabletest.language
+
+import com.intellij.lexer.FlexAdapter
+
+class TableTestLexerAdapter : FlexAdapter(TableTestLexer(null))