+++ src/main/kotlin/io/github/nchaugen/tabletest/language/TableTestSyntaxHighlighterFactory.kt
@@ -1,0 +1,12 @@
+package io.github.nchaugen.tabletest.language
+
+import com.intellij.openapi.fileTypes.SyntaxHighlighter
+import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFile
+
+internal class TableTestSyntaxHighlighterFactory : SyntaxHighlighterFactory() {
+
+    override fun getSyntaxHighlighter(project: Project?, virtualFile: VirtualFile?): SyntaxHighlighter =
+        TableTestSyntaxHighlighter()
+}